feat(setting): Refactor SettingService, fix APP_DEBUG=true APP_ENV=local APP_FAKER_LOCALE=id_ID APP_FALLBACK_LOCALE=id APP_KEY=base64:JmNh67Nn+cVrCcPIycspjxOj6tm4OWy19f577xZf7VI= APP_LOCALE=id APP_MAINTENANCE_DRIVER=file APP_NAME=Internara APP_TIMEZONE=Asia/Jakarta APP_URL=http://localhost AWS_ACCESS_KEY_ID= AWS_BUCKET= AWS_DEFAULT_REGION=us-east-1 AWS_SECRET_ACCESS_KEY= AWS_USE_PATH_STYLE_ENDPOINT=false BASH=/usr/bin/bash BASHOPTS=checkwinsize:cmdhist:complete_fullquote:extquote:force_fignore:globasciiranges:globskipdots:hostcomplete:interactive_comments:patsub_replacement:progcomp:sourcepath BASH_ALIASES=() BASH_ARGC=() BASH_ARGV=() BASH_CMDS=() BASH_EXECUTION_STRING='shopt -u promptvars nullglob extglob nocaseglob dotglob; { git commit -m "feat(setting): Refactor SettingService, fix `set` method array handling, and enhance test coverage"; }; __code=$?; pgrep -g 0 >/tmp/shell_pgrep_209eb8041c40.tmp 2>&1; exit $__code;' BASH_LINENO=() BASH_LOADABLES_PATH=/usr/local/lib/bash:/usr/lib/bash:/opt/local/lib/bash:/usr/pkg/lib/bash:/opt/pkg/lib/bash:. BASH_SOURCE=() BASH_VERSINFO=([0]="5" [1]="2" [2]="21" [3]="1" [4]="release" [5]="x86_64-pc-linux-gnu") BASH_VERSION='5.2.21(1)-release' BCRYPT_ROUNDS=12 BROADCAST_CONNECTION=log BUN_INSTALL=/home/reasnovynt/.bun CACHE_STORE=database COLORTERM=truecolor DBUS_SESSION_BUS_ADDRESS=unix:path=/run/user/1000/bus DB_CONNECTION=sqlite DEBUGINFOD_URLS='https://debuginfod.ubuntu.com ' DESKTOP_SESSION=ubuntu DIRSTACK=() DISPLAY=:0 EUID=1000 FILESYSTEM_DISK=local GDMSESSION=ubuntu GEMINI_CLI=1 GEMINI_CLI_NO_RELAUNCH=true GIT_PAGER=cat GNOME_DESKTOP_SESSION_ID=this-is-deprecated GNOME_KEYRING_CONTROL=/run/user/1000/keyring GNOME_SETUP_DISPLAY=:1 GNOME_SHELL_SESSION_MODE=ubuntu GNOME_TERMINAL_SCREEN=/org/gnome/Terminal/screen/b7e16005_9334_46e4_81e8_2910cf6165e8 GNOME_TERMINAL_SERVICE=:1.96 GROUPS=() GSM_SKIP_SSH_AGENT_WORKAROUND=true GTK_MODULES=gail:atk-bridge HISTFILESIZE=2147450879 HISTSIZE=2147450879 HOME=/home/reasnovynt HOSTNAME=code HOSTTYPE=x86_64 IFS=$' \t\n' IM_CONFIG_PHASE=1 LANG=en_US.UTF-8 LANGUAGE=en_US:en LC_ADDRESS=en_US.UTF-8 LC_CTYPE=en_US.UTF-8 LC_IDENTIFICATION=en_US.UTF-8 LC_MEASUREMENT=en_US.UTF-8 LC_MONETARY=en_US.UTF-8 LC_NAME=en_US.UTF-8 LC_NUMERIC=en_US.UTF-8 LC_PAPER=en_US.UTF-8 LC_TELEPHONE=en_US.UTF-8 LC_TIME=en_US.UTF-8 LESS=-R LESSCLOSE='/usr/bin/lesspipe %s %s' LESSOPEN='| /usr/bin/lesspipe %s' LIBVA_DRIVER_NAME=i965 LOGNAME=reasnovynt LOG_CHANNEL=stack LOG_DEPRECATIONS_CHANNEL=null LOG_LEVEL=debug LOG_STACK=single LSCOLORS=Gxfxcxdxdxegedabagacad LS_COLORS='rs=0:di=01;34:ln=01;36:mh=00:pi=40;33:so=01;35:do=01;35:bd=40;33;01:cd=40;33;01:or=40;31;01:mi=00:su=37;41:sg=30;43:ca=00:tw=30;42:ow=34;42:st=37;44:ex=01;32:*.tar=01;31:*.tgz=01;31:*.arc=01;31:*.arj=01;31:*.taz=01;31:*.lha=01;31:*.lz4=01;31:*.lzh=01;31:*.lzma=01;31:*.tlz=01;31:*.txz=01;31:*.tzo=01;31:*.t7z=01;31:*.zip=01;31:*.z=01;31:*.dz=01;31:*.gz=01;31:*.lrz=01;31:*.lz=01;31:*.lzo=01;31:*.xz=01;31:*.zst=01;31:*.tzst=01;31:*.bz2=01;31:*.bz=01;31:*.tbz=01;31:*.tbz2=01;31:*.tz=01;31:*.deb=01;31:*.rpm=01;31:*.jar=01;31:*.war=01;31:*.ear=01;31:*.sar=01;31:*.rar=01;31:*.alz=01;31:*.ace=01;31:*.zoo=01;31:*.cpio=01;31:*.7z=01;31:*.rz=01;31:*.cab=01;31:*.wim=01;31:*.swm=01;31:*.dwm=01;31:*.esd=01;31:*.avif=01;35:*.jpg=01;35:*.jpeg=01;35:*.mjpg=01;35:*.mjpeg=01;35:*.gif=01;35:*.bmp=01;35:*.pbm=01;35:*.pgm=01;35:*.ppm=01;35:*.tga=01;35:*.xbm=01;35:*.xpm=01;35:*.tif=01;35:*.tiff=01;35:*.png=01;35:*.svg=01;35:*.svgz=01;35:*.mng=01;35:*.pcx=01;35:*.mov=01;35:*.mpg=01;35:*.mpeg=01;35:*.m2v=01;35:*.mkv=01;35:*.webm=01;35:*.webp=01;35:*.ogm=01;35:*.mp4=01;35:*.m4v=01;35:*.mp4v=01;35:*.vob=01;35:*.qt=01;35:*.nuv=01;35:*.wmv=01;35:*.asf=01;35:*.rm=01;35:*.rmvb=01;35:*.flc=01;35:*.avi=01;35:*.fli=01;35:*.flv=01;35:*.gl=01;35:*.dl=01;35:*.xcf=01;35:*.xwd=01;35:*.yuv=01;35:*.cgm=01;35:*.emf=01;35:*.ogv=01;35:*.ogx=01;35:*.aac=00;36:*.au=00;36:*.flac=00;36:*.m4a=00;36:*.mid=00;36:*.midi=00;36:*.mka=00;36:*.mp3=00;36:*.mpc=00;36:*.ogg=00;36:*.ra=00;36:*.wav=00;36:*.oga=00;36:*.opus=00;36:*.spx=00;36:*.xspf=00;36:*~=00;90:*#=00;90:*.bak=00;90:*.crdownload=00;90:*.dpkg-dist=00;90:*.dpkg-new=00;90:*.dpkg-old=00;90:*.dpkg-tmp=00;90:*.old=00;90:*.orig=00;90:*.part=00;90:*.rej=00;90:*.rpmnew=00;90:*.rpmorig=00;90:*.rpmsave=00;90:*.swp=00;90:*.tmp=00;90:*.ucf-dist=00;90:*.ucf-new=00;90:*.ucf-old=00;90:' MACHTYPE=x86_64-pc-linux-gnu MAIL_FROM_ADDRESS=hello@example.com MAIL_FROM_NAME='${APP_NAME}' MAIL_HOST=127.0.0.1 MAIL_MAILER=log MAIL_PASSWORD=null MAIL_PORT=2525 MAIL_SCHEME=null MAIL_USERNAME=null MEMCACHED_HOST=127.0.0.1 MEMORY_PRESSURE_WATCH=/sys/fs/cgroup/user.slice/user-1000.slice/user@1000.service/session.slice/org.gnome.SettingsDaemon.MediaKeys.service/memory.pressure MEMORY_PRESSURE_WRITE=c29tZSAyMDAwMDAgMjAwMDAwMAA= NVM_BIN=/home/reasnovynt/.nvm/versions/node/v24.11.1/bin NVM_CD_FLAGS= NVM_DIR=/home/reasnovynt/.nvm NVM_INC=/home/reasnovynt/.nvm/versions/node/v24.11.1/include/node OLDPWD=/home/reasnovynt OPTERR=1 OPTIND=1 OSH=/home/reasnovynt/.oh-my-bash OSTYPE=linux-gnu PAGER=cat PAPERSIZE=letter PATH=/home/reasnovynt/.local/share/pnpm:/home/reasnovynt/.config/herd-lite/bin:/home/reasnovynt/.nvm/versions/node/v24.11.1/bin:/home/reasnovynt/.config/composer/vendor/bin:/home/reasnovynt/.bun/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin:/snap/bin PHP_INI_SCAN_DIR=/home/reasnovynt/.config/herd-lite/bin: PIPESTATUS=([0]="0") PNPM_HOME=/home/reasnovynt/.local/share/pnpm PPID=8811 PS4='+ ' PWD=/home/reasnovynt/Projects/reasnov/internara QT_ACCESSIBILITY=1 QT_IM_MODULE=ibus QUEUE_CONNECTION=database REDIS_CLIENT=phpredis REDIS_HOST=127.0.0.1 REDIS_PASSWORD=null REDIS_PORT=6379 SESSION_DOMAIN=null SESSION_DRIVER=database SESSION_ENCRYPT=false SESSION_LIFETIME=120 SESSION_MANAGER=local/code:@/tmp/.ICE-unix/2224,unix/code:/tmp/.ICE-unix/2224 SESSION_PATH=/ SHELL=/bin/bash SHELLOPTS=braceexpand:hashall:interactive-comments SHLVL=2 SSH_AUTH_SOCK=/run/user/1000/keyring/ssh SYSTEMD_EXEC_PID=2445 TERM=xterm UID=1000 USER=reasnovynt USERNAME=reasnovynt VITE_APP_NAME='${APP_NAME}' VTE_VERSION=7600 WAYLAND_DISPLAY=wayland-0 XAUTHORITY=/run/user/1000/.mutter-Xwaylandauth.J1M8I3 XDG_CONFIG_DIRS=/etc/xdg/xdg-ubuntu:/etc/xdg XDG_CURRENT_DESKTOP=ubuntu:GNOME XDG_DATA_DIRS=/usr/share/ubuntu:/usr/share/gnome:/home/reasnovynt/.local/share/flatpak/exports/share:/var/lib/flatpak/exports/share:/usr/local/share/:/usr/share/:/var/lib/snapd/desktop XDG_MENU_PREFIX=gnome- XDG_RUNTIME_DIR=/run/user/1000 XDG_SESSION_CLASS=user XDG_SESSION_DESKTOP=ubuntu XDG_SESSION_TYPE=wayland XMODIFIERS=@im=ibus _=dotglob method array handling, and enhance test coverage

+++ modules/Setting/src/Services/SettingService.php
@@ -2,12 +2,43 @@
 
 namespace Modules\Setting\Services;
 
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Cache;
 use Modules\Setting\Contracts\Services\SettingService as SettingServiceContract;
 use Modules\Setting\Models\Setting;
 
 class SettingService implements SettingServiceContract
 {
+    /**
+     * The Setting model instance.
+     *
+     * @var \Modules\Setting\Models\Setting
+     */
+    protected Setting $model;
+
+    /**
+     * Create a new SettingService instance.
+     *
+     * @param  \Modules\Setting\Models\Setting  $model The Setting model instance.
+     */
+    public function __construct(Setting $model)
+    {
+        $this->model = $model;
+    }
+
+    /**
+     * Get all settings.
+     *
+     * @param  array  $columns  The columns to select.
+     * @param  bool  $skipCache  Whether to bypass the cache and fetch directly from the database.
+     * @return \Illuminate\Support\Collection A collection of all Setting models.
+     */
+    public function all(array $columns = ['*'], bool $skipCache = false): Collection
+    {
+        return $this->remember('settings.all', fn () => $this->model->all($columns), $skipCache);
+    }
+
     /**
      * {@inheritDoc}
      */
@@ -22,19 +53,15 @@ class SettingService implements SettingServiceContract
             return $results;
         }
 
-        if ($skipCached) {
-            $this->forgetKey($key);
-        }
-
-        return $this->rememberKey($key, function () use ($key, $default) {
-            $setting = Setting::find($key);
+        return $this->remember('settings.'.$key, function () use ($key, $default) {
+            $setting = $this->model->find($key);
 
             if (! $setting) {
                 return $default;
             }
 
             return $setting->value;
-        });
+        }, $skipCached);
     }
 
     /**
@@ -45,7 +72,9 @@ class SettingService implements SettingServiceContract
         if (is_array($key)) {
             $success = true;
             foreach ($key as $k => $v) {
-                if (! $this->set($k, $v, $extraAttributes[$k] ?? [])) {
+                $value = $v['value'] ?? $v;
+                $extraAttributes = is_array($v) ? array_diff_key($v, ['value' => null]) : [];
+                if (! $this->set($k, $value, $extraAttributes)) {
                     $success = false;
                 }
             }
@@ -53,17 +82,15 @@ class SettingService implements SettingServiceContract
             return $success;
         }
 
-        // The cast handles value stringification and type attribute setting.
-        // We only pass the raw value and any other extra attributes.
-        $setting = Setting::updateOrCreate(
+        $setting = $this->model->updateOrCreate(
             ['key' => $key],
             array_merge(
-                ['value' => $value], // Pass the raw value, cast will process it
+                ['value' => $value],
                 $extraAttributes
             )
         );
 
-        $this->forgetKey($key, $setting->group ?? null);
+        $this->clearCache($key, $setting->group);
 
         return (bool) $setting;
     }
@@ -71,40 +98,74 @@ class SettingService implements SettingServiceContract
     /**
      * {@inheritDoc}
      */
-    public function getByGroup(string $name): \Illuminate\Support\Collection
+    public function getByGroup(string $name, bool $skipCache = false): Collection
     {
-        return $this->rememberKey($name, fn () => Setting::group($name)->get(), isGroup: true);
+        return $this->remember('settings.group.'.$name, fn () => $this->model->group($name)->get(), $skipCache);
     }
 
+    /**
+     * {@inheritDoc}
+     */
     public function delete(string $key): bool
     {
-        $setting = Setting::findOrFail($key);
+        $setting = $this->model->find($key);
+
+        if (! $setting) {
+            return false;
+        }
 
         if ($deleted = $setting->delete()) {
-            $this->forgetKey($key);
+            $this->clearCache($key, $setting->group);
         }
 
         return $deleted;
     }
 
-    private function rememberKey(string $key, mixed $callback = null, bool $isGroup = false): mixed
+    /**
+     * {@inheritDoc}
+     */
+    public function query(array $columns = ['*']): Builder
     {
-        $cacheKey = $isGroup ? "settings.group.{$key}" : "settings.{$key}";
-
-        return Cache::remember($cacheKey, now()->addDay(), function () use ($callback) {
-            if (is_callable($callback)) {
-                return $callback();
-            }
-
-            return $callback;
-        });
+        return $this->model->query()->select($columns);
     }
 
-    private function forgetKey(string $key, ?string $group = null): void
+    /**
+     * Retrieve an item from the cache or store the result of a callback forever.
+     *
+     * @param string $cacheKey The key for the cache item.
+     * @param callable $callback The callback to execute if the item is not in the cache.
+     * @param bool $skipCache Whether to skip the cache and execute the callback directly.
+     * @return mixed
+     */
+    protected function remember(string $cacheKey, callable $callback, bool $skipCache = false): mixed
+    {
+        if ($skipCache) {
+            Cache::forget($cacheKey);
+
+            return $callback();
+        }
+
+        return Cache::remember($cacheKey, now()->addDay(), $callback);
+    }
+
+    /**
+     * Clear the cache for a specific setting and its related groups.
+     *
+     * @param string $key The key of the setting whose cache to clear.
+     * @param string|null $group The group name the setting belongs to, if any.
+     * @return void
+     */
+    protected function clearCache(string $key, ?string $group = null): void
     {
         // Clear cache for this specific setting
         Cache::forget('settings.'.$key);
-        // Invalidate cache for any group this setting might belong to
-        Cache::forget('settings.group.'.($group ?? ''));
+
+        // Clear cache for the group this setting belongs to, if any
+        if ($group) {
+            Cache::forget('settings.group.'.$group);
+        }
+
+        // Clear cache for all settings, as this setting might affect the 'all' collection
+        Cache::forget('settings.all');
     }
 }
